+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     - Added matrix mapping parameters on rows and columns for use with midi-mapping
     - Added option for excluding attenuverters from scenes
     - Added ability to copy scenes
+    - Added ability to reset scenes
     - Fixed fading if either fade-in or fade-out is set to zero
 - Added dark panels for all modules (#15)
 - Added globals settings, esp. for dark panels als default